Add files via upload
Fly like a bird! Then hit a wall.

The Essential: Jetpacks Mod delivers highly augmented armour granting the capability of flight, all through the convenience of a back item - like all traditional jetpacks should be. There are 3 jetpacks, each with their own unique flight mode. Additionally, there is armour with functionality engineered toward benefiting the wearer during flight. The weapons, and items can be used without the need for a jetpack, however they have been carefully designed to work better in flight - with the use of a Jetpack.

Experiment with complex combinations of Augments, weapons and techs, either from this mod, another mod or the vanilla base game content with your jetpack of choice. The sky is.... sort of the limit, but you get the analogy.

Polishing this mod has been great fun. Aside from bug fixes as I believe the focus should be on the Jetpacks themselves.

I suggest taking a look at the Mod's Readme which is in the form of a hand-craftable codex.
In order to build the Avionics lab, you can do so at a Tier 1 Inventor's Table.

I hope ye all enjoy this mod as much as I enjoyed makin it me hearties!

checkout my discord server: https://discord.gg/m2cvz2g
If you want to contribute or share any ideas you can do so.